+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+  * WPTrackserver and Directory allow renaming by assinging new id_in_backend
+
 1.6.0 release 2019-01-xx
 ------------------------